# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar Trust Escrow Backend API
  version: 0.1.0
  description: Shared pagination contract for collection endpoints.
paths:
  /api/v1/escrows:
    get:
      summary: List escrows
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Standardized paginated escrow list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /api/v1/escrows/{id}/milestones:
    get:
      summary: List milestones for an escrow
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Standardized paginated milestone list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /api/v1/users/{address}/escrows:
    get:
      summary: List escrows for a user
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Standardized paginated user escrow list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /api/v1/disputes:
    get:
      summary: List disputes
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Standardized paginated dispute list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /api/v1/reputation/leaderboard:
    get:
      summary: List reputation leaderboard entries
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Standardized paginated leaderboard response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
components:
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number. Values below 1 are normalized to 1.
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Page size. Defaults to 20 and is capped at 100.
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  schemas:
    PaginatedResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
        - totalPages
        - hasNextPage
        - hasPreviousPage
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false